Remove non-essential programs from runtime/
