+++ Cargo.toml
@@ -31,6 +31,7 @@ solana = { path = "core", version = "0.13.0" }
 solana-logger = { path = "logger", version = "0.13.0" }
 solana-netutil = { path = "netutil", version = "0.13.0" }
 solana-runtime = { path = "runtime", version = "0.13.0" }
+solana-budget-program = { path = "programs/budget", version = "0.13.0" }
 solana-sdk = { path = "sdk", version = "0.13.0" }
 sys-info = "0.5.6"
