+++ runtime/Cargo.toml
@@ -22,12 +22,9 @@ rand = "0.6.5"
 serde = "1.0.88"
 serde_derive = "1.0.88"
 serde_json = "1.0.38"
-solana-budget-api = { path = "../programs/budget_api", version = "0.13.0" }
 solana-logger = { path = "../logger", version = "0.13.0" }
 solana-metrics = { path = "../metrics", version = "0.13.0" }
 solana-sdk = { path = "../sdk", version = "0.13.0" }
-solana-storage-api = { path = "../programs/storage_api", version = "0.13.0" }
-solana-token-api = { path = "../programs/token_api", version = "0.13.0" }
 solana-vote-api = { path = "../programs/vote_api", version = "0.13.0" }
 
 [lib]
