+++ sdk/src/genesis_block.rs
@@ -24,6 +24,7 @@ pub struct GenesisBlock {
     pub slots_per_epoch: u64,
     pub stakers_slot_offset: u64,
     pub epoch_warmup: bool,
+    pub native_programs: Vec<(String, Pubkey)>,
 }
 
 impl GenesisBlock {
@@ -57,6 +58,7 @@ impl GenesisBlock {
                 slots_per_epoch: DEFAULT_SLOTS_PER_EPOCH,
                 stakers_slot_offset: DEFAULT_SLOTS_PER_EPOCH,
                 epoch_warmup: true,
+                native_programs: vec![],
             },
             mint_keypair,
         )
