+++ wallet/Cargo.toml
@@ -24,5 +24,8 @@ solana-sdk = { path = "../sdk", version = "0.13.0" }
 solana-vote-api = { path = "../programs/vote_api", version = "0.13.0" }
 solana-vote-signer = { path = "../vote-signer", version = "0.13.0" }
 
+[dev-dependencies]
+solana-budget-program = { path = "../programs/budget", version = "0.13.0" }
+
 [features]
 cuda = ["solana/cuda"]
